attn_res: move the model code under deepseek_v3_d_p
Aligns with how KDA is organized so both Kimi K3 features land in one tree:
the torch references under `reference/attn_res/`, the ttnn composite under
`tt/attn_res/`, and the tests under `tests/attn_res/` split model vs perf.

The model PCC gate moves out of the ttnn nightly op tree with them — it drives
`TtAttnRes`, not an op — so the Blackhole e2e manifest row follows its path.
